+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.8.1] - 2018-XX-XX
+### Changed
+- [Credits] Lover of the month
+
 ## [0.8.0] - 2018-06-29
 ### Added
 - Push Notifications